requirements.txt
Django==5.2.8
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
sqlparse==0.5.1
tzdata==2024.2
asgiref==3.8.1
requests==2.31.0
